Implement pointer tagging for 'big' families #14373
Formerly we punted on these and evaluated constructors always
got a tag of 1.

We now cascade switches because we have to check the tag first
and when it is MAX_PTR_TAG then get the precise tag from the info
table and switch on that. The only technically tricky part is that the
default case needs (logical) duplication. To do this we emit an extra
label for it and branch to that from the second switch. This avoids
duplicated codegen.